// src/app/shared/interfaces/response.ts
export interface ResponseDto {  
    summary: string;
    stressUnitOfMeasure: string;
    stressScore: string;
    depressionDescription: string;
    depressionUnitOfMeasure: string;
    anxietyDescription: string;
    description: string;
    stressDescription: string;
    anxietyScore: string;
    depressionScore: string;
    depressionLabel: string;
    stressLabel: string;
    takenOnDateTime: string;
    anxietyLabel: string;
    anxietyUnitOfMeasure: string;
};